name changes to some methods

+++ src/entity/CommonUser.java
@@ -36,7 +36,7 @@ public class CommonUser implements User{
 
     private String paceType;
 
-    private int requiredCalories;
+    private double requiredCalories;
 
 
 
@@ -482,7 +482,7 @@ public class CommonUser implements User{
 
 
 
-    public int getRequiredCalories() {
+    public double getRequiredCalories() {
         return requiredCalories;
     }
 
@@ -537,7 +537,7 @@ public class CommonUser implements User{
     }
 
     @Override
-    public void setRequiredCalories(int requiredCalories) {
+    public void setRequiredCalories(double requiredCalories) {
         this.requiredCalories = requiredCalories;
     }
